# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  gosh_internal_dto_request.CreateBannerRequest:
    properties:
      background:
        maxLength: 128
        type: string
      description:
        maxLength: 256
        type: string
      image:
        maxLength: 256
        type: string
      link:
        maxLength: 256
        type: string
      sort_order:
        type: integer
      subtitle:
        maxLength: 128
        type: string
      title:
        maxLength: 64
        type: string
    type: object
  gosh_internal_dto_request.CreateCategoryRequest:
    properties:
      banner:
        maxLength: 256
        type: string
      icon:
        maxLength: 32
        type: string
      name:
        maxLength: 32
        type: string
      parent_id:
        type: integer
      sort_order:
        type: integer
    required:
    - name
    type: object
...
  gosh_internal_dto_request.CreateReviewRequest:
    properties:
      content:
        maxLength: 500
        type: string
      images:
        type: string
      product_id:
        type: integer
      score:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - product_id
    - score
    type: object
...
  gosh_internal_dto_request.UpdateBannerRequest:
    properties:
      background:
        maxLength: 128
        type: string
      description:
        maxLength: 256
        type: string
      image:
        maxLength: 256
        type: string
      link:
        maxLength: 256
        type: string
      sort_order:
        type: integer
      status:
        enum:
        - "on"
        - "off"
        type: string
      subtitle:
        maxLength: 128
        type: string
      title:
        maxLength: 64
        type: string
    type: object
  gosh_internal_dto_request.UpdateBrandStoryRequest:
    properties:
      badge:
        maxLength: 32
        type: string
      description:
        maxLength: 256
        type: string
      link:
        maxLength: 256
        type: string
      logo:
        maxLength: 256
        type: string
      title:
        maxLength: 64
        type: string
    type: object
...
  gosh_internal_dto_request.UpdateCategoryRequest:
    properties:
      banner:
        maxLength: 256
        type: string
      icon:
        maxLength: 32
        type: string
      name:
        maxLength: 32
        type: string
      sort_order:
        type: integer
    type: object
...
  gosh_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
host: localhost:9292
info:
  contact:
    email: support@gosh-mall.com
    name: API Support
  description: Gosh Mall is a full-stack e-commerce platform with user management,
    product catalog, shopping cart, order processing, payment, coupons, and admin
    features.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Gosh Mall API
  version: "1.0"
paths:
...
  /admin/banners:
    get:
      consumes:
      - application/json
      description: Admin lists all banners
      parameters:
      - description: Status filter
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: List all banners
      tags:
      - Banners
    post:
      consumes:
      - application/json
      description: Admin creates a new banner
      parameters:
      - description: Banner info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gosh_internal_dto_request.CreateBannerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create a banner
      tags:
      - Banners
  /admin/banners/{id}:
    delete:
      consumes:
      - application/json
      description: Admin deletes a banner
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a banner
      tags:
      - Banners
    put:
      consumes:
      - application/json
      description: Admin updates a banner
      parameters:
      - description: Banner ID
        in: path
        name: id
        required: true
        type: integer
      - description: Banner update info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gosh_internal_dto_request.UpdateBannerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update a banner
      tags:
      - Banners
  /admin/brand-story:
    put:
      consumes:
      - application/json
      description: Admin updates the brand story
      parameters:
      - description: Brand story info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gosh_internal_dto_request.UpdateBrandStoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update brand story
      tags:
      - Banners
  /admin/categories:
    post:
      consumes:
      - application/json
      description: Create a new category
      parameters:
      - description: Category info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gosh_internal_dto_request.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create a category
      tags:
      - Categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Delete a category
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category update info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gosh_internal_dto_request.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Update a category
      tags:
      - Categories
...
  /banners:
    get:
      consumes:
      - application/json
      description: Get all active banners for public display
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      summary: Get active banners
      tags:
      - Banners
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get hierarchical category tree
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      summary: Get category tree
      tags:
      - Categories
  /categories/{id}:
    get:
      consumes:
      - application/json
      description: Get category info by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      summary: Get category by ID
      tags:
      - Categories
...
  /reviews:
    get:
      consumes:
      - application/json
      description: List reviews for a product with pagination
      parameters:
      - description: Product ID
        in: query
        name: product_id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      summary: List reviews
      tags:
      - Reviews
    post:
      consumes:
      - application/json
      description: Create a product review
      parameters:
      - description: Review info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/gosh_internal_dto_request.CreateReviewRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/gosh_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - Reviews
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"